Revert "Add tests"
