--- JQuiz/src/test/java/question/TestAddManyQuestions.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestAddManyQuestions {
-}
