--- JQuiz/src/test/java/question/TestAddOneQuestion.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestAddOneQuestion {
-}
