--- JQuiz/src/test/java/question/TestCountOfRandomQuestionsAsAll.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestCountOfRandomQuestionsAsAll {
-}
