--- JQuiz/src/test/java/question/TestCountOfRandomQuestionsLessThanAll.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestCountOfRandomQuestionsLessThanAll {
-}
