--- JQuiz/src/test/java/question/TestCountOfRandomQuestionsMoreThanAll.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestCountOfRandomQuestionsMoreThanAll {
-}
