--- JQuiz/src/test/java/question/TestDeleteQuestion.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestDeleteQuestion {
-}
