--- JQuiz/src/test/java/question/TestFindExistingQuestion.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestFindExistingQuestion {
-}
