--- JQuiz/src/test/java/question/TestFindNotExistingQuestion.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestFindNotExistingQuestion {
-}
