--- JQuiz/src/test/java/question/TestIsDifferentRandomQuestions.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestIsDifferentRandomQuestions {
-}
