--- JQuiz/src/test/java/question/TestUpdateQuestion.java
@@ -1,4 +1,0 @@
-package question;
-
-public class TestUpdateQuestion {
-}
