--- JQuiz/src/test/java/result/TestAddManyResults.java
@@ -1,4 +1,0 @@
-package result;
-
-public class TestAddManyResults {
-}
