--- JQuiz/src/test/java/result/TestAddOneResult.java
@@ -1,4 +1,0 @@
-package result;
-
-public class TestAddOneResult {
-}
