--- JQuiz/src/test/java/result/TestDeleteResult.java
@@ -1,4 +1,0 @@
-package result;
-
-public class TestDeleteResult {
-}
