--- JQuiz/src/test/java/result/TestFindExistingResult.java
@@ -1,4 +1,0 @@
-package result;
-
-public class TestFindExistingResult {
-}
