--- JQuiz/src/test/java/result/TestFindNotExistingResult.java
@@ -1,4 +1,0 @@
-package result;
-
-public class TestFindNotExistingResult {
-}
