--- JQuiz/src/test/java/result/TestUpdateResult.java
@@ -1,4 +1,0 @@
-package result;
-
-public class TestUpdateResult {
-}
