--- JQuiz/src/test/java/user/TestAddManyUsers.java
@@ -1,4 +1,0 @@
-package user;
-
-public class TestAddManyUsers {
-}
