--- JQuiz/src/test/java/user/TestDeleteUser.java
@@ -1,4 +1,0 @@
-package user;
-
-public class TestDeleteUser {
-}
