--- JQuiz/src/test/java/user/TestFindExistingUser.java
@@ -1,4 +1,0 @@
-package user;
-
-public class TestFindExistingUser {
-}
