--- JQuiz/src/test/java/user/TestFindNotExistingUser.java
@@ -1,4 +1,0 @@
-package user;
-
-public class TestFindNotExistingUser {
-}
