--- JQuiz/src/test/java/user/TestUpdateUser.java
@@ -1,4 +1,0 @@
-package user;
-
-public class TestUpdateUser {
-}
